requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
gspread>=6.0.0
google-auth>=2.0.0
toml>=0.10.0
